Actualizacion de paquetes
realizacion de crud en administrador y mostrar paquetes en el index para realizar compra

+++ src/main/java/com/viajesglobal/repository/PaqueteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PaqueteRepository extends JpaRepository<Paquete, Long> {
+public interface PaqueteRepository extends JpaRepository<Paquete, Integer> {
 }
